Research_code
These are the files I use for my minor thesis on remote sensing and image processing.

The aim of my research is to create a higher resolution land cover crop map of Victoria state in Australia. For which I'll be using Google Earth Engine to pull data and image pre-processing.